Refactoring: Abstract notification send service using interface
